requirements.txt
requests
beautifulsoup4
sqlite-utils
fake-useragent
discord.py
telethon
transformers
torch
tenacity
robotexclusionrulesparser
APScheduler
python-telegram-bot
python-dotenv
pydantic
aiosqlite
